requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    make freeze
#
astroid==2.8.0
    # via whispers (setup.py)
beautifulsoup4==4.10.0
    # via whispers (setup.py)
jproperties==2.1.1
    # via whispers (setup.py)
lazy-object-proxy==1.6.0
    # via astroid
luhn==0.2.0
    # via whispers (setup.py)
lxml==4.6.3
    # via whispers (setup.py)
python-levenshtein==0.12.2
    # via whispers (setup.py)
pyyaml==5.4.1
    # via whispers (setup.py)
six==1.16.0
    # via jproperties
soupsieve==2.2.1
    # via beautifulsoup4
typing-extensions==3.10.0.2
    # via astroid
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
